requirements.txt
cdsapi==0.6.1
importlib_metadata==6.6.0
livelossplot==0.5.5
matplotlib==3.7.1
numpy==1.24.1
pandas==1.5.3
Pillow==10.0.0
pyaerocom==0.14.0
pytest==7.1.2
Requests==2.31.0
scipy==1.10.0
tomli==2.0.1
torch==2.0.1
tqdm==4.64.1
xarray==2022.10.0
scikit-image
protobuf==3.9.2
tabulate
sphinx
sphinx_rtd_theme
